refactor(ui): comment png imports, svg ok, needs to cover export alias and destructured
