+++ packages/twenty-ui/jest.config.ts
@@ -21,7 +21,7 @@ const jestConfig: JestConfigWithTsJest = {
     ],
   },
   moduleNameMapper: {
-    '\\.(jpg|jpeg|png|gif|webp|svg|svg\\?react)$':
+    '\\.(jpg|jpeg|png|gif|webp|svg|svg\\)$':
       '<rootDir>/__mocks__/imageMock.js',
     ...pathsToModuleNameMapper(tsConfig.compilerOptions.paths, {
       prefix: '<rootDir>/',
