+++ packages/twenty-ui/package.json
@@ -11,7 +11,9 @@
     "@types/babel__preset-env": "^7",
     "@types/react": "^18.2.39",
     "@types/react-dom": "^18.2.15",
+    "babel-plugin-inline-react-svg": "^2.0.2",
     "babel-plugin-module-resolver": "^5.0.2",
+    "cpx": "^1.5.0",
     "glob": "^11.0.1",
     "tsx": "^4.19.3"
   },
@@ -19,23 +21,24 @@
     "@emotion/is-prop-valid": "^1.3.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
+    "@linaria/react": "^6.2.1",
     "@monaco-editor/react": "^4.6.0",
     "@sniptt/guards": "^0.2.0",
     "date-fns": "^2.30.0",
     "framer-motion": "^11.18.0",
+    "hex-rgb": "^5.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-responsive": "^9.0.2",
     "react-router-dom": "^6.4.4",
     "react-tooltip": "^5.13.1",
     "recoil": "^0.7.7",
-    "hex-rgb": "^5.0.0",
     "twenty-shared": "workspace:*",
     "zod": "3.23.8",
-    "@linaria/react": "^6.2.1"
+    "@tabler/icons-react": "^3.31.0"
   },
   "scripts": {
-    "build": "preconstruct build"
+    "build": "preconstruct build && cpx \"assets/**/*\" dist/assets"
   },
   "preconstruct": {
     "entrypoints": [
