+++ packages/twenty-ui/src/display/icon/components/IconGoogle.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 
-import IconGoogleRaw from '@assets/icons/google.svg?react';
+import IconGoogleRaw from '@assets/icons/google.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IconGoogleProps = Pick<IconComponentProps, 'size'>;
