+++ packages/twenty-ui/src/display/icon/components/IconGoogleCalendar.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 
-import IconGoogleCalendarRaw from '@assets/icons/google-calendar.svg?react';
+import IconGoogleCalendarRaw from '@assets/icons/google-calendar.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IconGoogleCalendarProps = Pick<IconComponentProps, 'size'>;
