+++ packages/twenty-ui/src/display/icon/components/IconMicrosoft.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 
-import IconMicrosoftRaw from '@assets/icons/microsoft.svg?react';
+import IconMicrosoftRaw from '@assets/icons/microsoft.svg';
 
 interface IconMicrosoftProps {
   size?: number | string;
