+++ packages/twenty-ui/src/display/icon/components/IconMicrosoftCalendar.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 
-import IconMicrosoftCalendarRaw from '@assets/icons/microsoft-calendar.svg?react';
+import IconMicrosoftCalendarRaw from '@assets/icons/microsoft-calendar.svg';
 
 interface IconMicrosoftCalendarProps {
   size?: number | string;
