+++ packages/twenty-ui/src/display/icon/components/IconMicrosoftOutlook.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 
-import IconMicrosoftOutlookRaw from '@assets/icons/microsoft-outlook.svg?react';
+import IconMicrosoftOutlookRaw from '@assets/icons/microsoft-outlook.svg';
 
 interface IconMicrosoftOutlookProps {
   size?: number | string;
