+++ packages/twenty-ui/src/display/icon/components/IconTwentyStar.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react';
 
-import IconTwentyStarRaw from '@assets/icons/twenty-star.svg?react';
+import IconTwentyStarRaw from '@assets/icons/twenty-star.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IconTwentyStarProps = Pick<IconComponentProps, 'size' | 'stroke'>;
