+++ packages/twenty-ui/src/display/icon/components/IllustrationIconCalendarEvent.tsx
@@ -1,4 +1,4 @@
-import IllustrationIconCalendarEventRaw from '@assets/icons/illustration-calendar-event.svg?react';
+import IllustrationIconCalendarEventRaw from '@assets/icons/illustration-calendar-event.svg';
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
