+++ packages/twenty-ui/src/display/icon/components/IllustrationIconCalendarTime.tsx
@@ -1,4 +1,4 @@
-import IllustrationIconCalendarTimeRaw from '@assets/icons/illustration-calendar-time.svg?react';
+import IllustrationIconCalendarTimeRaw from '@assets/icons/illustration-calendar-time.svg';
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
