+++ packages/twenty-ui/src/display/icon/components/IllustrationIconCurrency.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 
-import IllustrationIconCurrencyRaw from '@assets/icons/illustration-currency.svg?react';
+import IllustrationIconCurrencyRaw from '@assets/icons/illustration-currency.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IllustrationIconCurrencyProps = Pick<IconComponentProps, 'size'>;
