+++ packages/twenty-ui/src/display/icon/components/IllustrationIconJson.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 
-import IllustrationIconJsonRaw from '@assets/icons/illustration-json.svg?react';
+import IllustrationIconJsonRaw from '@assets/icons/illustration-json.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IllustrationIconJsonProps = Pick<IconComponentProps, 'size'>;
