+++ packages/twenty-ui/src/display/icon/components/IllustrationIconMap.tsx
@@ -1,4 +1,4 @@
-import IllustrationIconMapRaw from '@assets/icons/illustration-map.svg?react';
+import IllustrationIconMapRaw from '@assets/icons/illustration-map.svg';
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
