+++ packages/twenty-ui/src/display/icon/components/IllustrationIconSetting.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 
-import IllustrationIconSettingRaw from '@assets/icons/illustration-setting.svg?react';
+import IllustrationIconSettingRaw from '@assets/icons/illustration-setting.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IllustrationIconSettingProps = Pick<IconComponentProps, 'size'>;
