+++ packages/twenty-ui/src/display/icon/components/IllustrationIconStar.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 
-import IllustrationIconStarRaw from '@assets/icons/illustration-star.svg?react';
+import IllustrationIconStarRaw from '@assets/icons/illustration-star.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IllustrationIconStarProps = Pick<IconComponentProps, 'size'>;
