+++ packages/twenty-ui/src/display/icon/components/IllustrationIconTag.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 
-import IllustrationIconTagRaw from '@assets/icons/illustration-tag.svg?react';
+import IllustrationIconTagRaw from '@assets/icons/illustration-tag.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IllustrationIconTagProps = Pick<IconComponentProps, 'size'>;
