+++ packages/twenty-ui/src/display/icon/components/IllustrationIconTags.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@emotion/react';
 import { IllustrationIconWrapper } from '@ui/display/icon/components/IllustrationIconWrapper';
 
-import IllustrationIconTagsRaw from '@assets/icons/illustration-tags.svg?react';
+import IllustrationIconTagsRaw from '@assets/icons/illustration-tags.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
 
 type IllustrationIconTagsProps = Pick<IconComponentProps, 'size'>;
