+++ packages/twenty-ui/src/images.d.ts
@@ -19,11 +19,6 @@ declare module "*.gif" {
 }
 
 declare module "*.svg" {
-  const value: string;
-  export default value;
-} 
-
-declare module "*.svg?react" {
   import * as React from "react";
 
   const ReactComponent: React.FunctionComponent<
