+++ packages/twenty-ui/src/input/components/IconListViewGrip.tsx
@@ -1,5 +1,5 @@
+import IconListViewGripRaw from '@assets/misc/list-view-grip.svg';
 import { IconComponentProps } from '@ui/display/icon/types/IconComponent';
-import IconListViewGripRaw from '@ui/input/components/list-view-grip.svg?react';
 
 type IconListViewGripProps = Pick<IconComponentProps, 'size' | 'stroke'>;
