+++ packages/twenty-ui/src/navigation/menu-item/components/MenuItemSelect.tsx
@@ -1,7 +1,8 @@
 import { css, useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 
-import { IconCheck, IconChevronRight, IconComponent } from '@ui/display';
+import { IconCheck, IconChevronRight } from '@tabler/icons-react';
+import { IconComponent } from '@ui/display';
 import { MenuItemLeftContent } from '../internals/components/MenuItemLeftContent';
 import { StyledMenuItemBase } from '../internals/components/StyledMenuItemBase';
